# contracts/pms-console-api/v1/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "PMS Console API Contract",
    "version": "1.0.0",
    "description": "Contract Frozen from Candidate 3. Transport adapter contract only; authentication, authorization, RBAC, login and sessions remain deferred. X-Actor-ID remains audit context."
  },
  "servers": [
    {
      "url": "/api/console/v1"
    }
  ],
...
  "paths": {
    "/provider-packages": {
      "get": {
        "operationId": "listProviderPackages",
        "tags": [
          "ProviderPackages"
        ],
        "summary": "listProviderPackages",
        "responses": {
          "200": {
            "description": "Page",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "items"
                  ],
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ProviderPackage"
                      }
                    },
                    "nextCursor": {
                      "type": "string"
                    }
                  },
                  "additionalProperties": false
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "parameters": [
          {
            "name": "providerType",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            }
          },
          {
            "name": "hostingMode",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "vendor_managed",
                "platform_managed"
              ]
            }
          },
          {
            "name": "componentStatus",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "passed",
                "partial",
                "pending",
                "failed"
              ]
            }
          },
          {
            "name": "realResourceStatus",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "qualified",
                "pending",
                "failed",
                "not_applicable"
              ]
            }
          },
          {
            "$ref": "#/components/parameters/CorrelationIdHeader"
          }
        ],
        "description": "Console transport projection of existing SMPP capability `listProviderPackages`."
      }
    },
    "/provider-packages/{packageId}": {
      "get": {
        "operationId": "getProviderPackage",
        "tags": [
          "ProviderPackages"
        ],
        "summary": "getProviderPackage",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderPackage"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "parameters": [
          {
            "name": "packageId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            }
          },
          {
            "name": "version",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            }
          },
          {
            "$ref": "#/components/parameters/CorrelationIdHeader"
          }
        ],
        "description": "Console transport projection of existing SMPP capability `getProviderPackage`."
      }
    },
...
  },
  "components": {
    "schemas": {
      "ProblemDetails": {
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "code"
        ],
        "properties": {
          "type": {
            "type": "string",
            "description": "Stable URN or URI identifying the problem category."
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "minimum": 400,
            "maximum": 599
          },
          "code": {
            "$ref": "#/components/schemas/ProblemCode"
          },
          "detail": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          },
          "correlationId": {
            "type": "string"
          }
        },
        "additionalProperties": false,
        "description": "Transport projection of existing SMPP public errors. No retryability policy is frozen in V1."
      },
      "ProblemCode": {
        "type": "string",
        "enum": [
          "INVALID_REQUEST",
          "INVALID_JSON",
          "REQUEST_BODY_TOO_LARGE",
          "INTERNAL_ERROR",
          "ROUTE_NOT_FOUND",
          "INVALID_IDENTIFIER",
          "INVALID_DOMAIN_VALUE",
          "INVALID_STATE_TRANSITION",
          "DUPLICATE_RESOURCE_BINDING",
          "RESOURCE_BINDING_NOT_FOUND",
          "ENTITY_ALREADY_EXISTS",
          "ENTITY_NOT_FOUND",
          "OPTIMISTIC_CONCURRENCY_CONFLICT",
          "LEASE_NOT_OWNED",
          "RUNTIME_DEPLOYMENT_NOT_FOUND",
          "RUNTIME_DEPLOYMENT_PROVIDER_UNAVAILABLE",
          "RUNTIME_DEPLOYMENT_CONFIG_PROFILE_UNAVAILABLE",
          "RUNTIME_DEPLOYMENT_DATABASE_PROFILE_UNAVAILABLE",
          "RUNTIME_DEPLOYMENT_REPLICA_COUNT_UNSUPPORTED",
          "RUNTIME_DEPLOYMENT_REVISION_CONFLICT",
          "RUNTIME_PROCESS_NOT_FOUND",
          "CONFIGURATION_DEFINITION_NOT_FOUND",
          "CONFIGURATION_TARGET_NOT_ALLOWED",
          "CONFIGURATION_BUSINESS_KEY_CONFLICT",
          "CONFIGURATION_DRAFT_NOT_FOUND",
          "CONFIGURATION_DRAFT_VERSION_CONFLICT",
          "CONFIGURATION_DRAFT_NOT_VALIDATED",
          "CONFIGURATION_PUBLISH_CONFLICT",
          "CONFIGURATION_REVISION_NOT_FOUND",
          "CONFIGURATION_ROLLBACK_TARGET_MISMATCH",
          "CONFIGURATION_INPUT_INVALID",
          "REGISTRY_SNAPSHOT_NOT_FOUND"
        ]
      },
...
      "ProviderPackage": {
        "type": "object",
        "required": [
          "packageId",
          "packageVersion",
          "providerType",
          "hostingModes",
          "configSchemaId",
          "compatibleRuntimeVersion",
          "protocolMode",
          "qualification"
        ],
        "properties": {
          "packageId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "packageVersion": {
            "type": "string",
            "minLength": 1,
            "maxLength": 64
          },
          "providerType": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "hostingModes": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "vendor_managed",
                "platform_managed"
              ]
            }
          },
          "configSchemaId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "compatibleRuntimeVersion": {
            "type": "string"
          },
          "protocolMode": {
            "type": "string"
          },
          "qualification": {
            "type": "object",
            "required": [
              "componentStatus",
              "realResourceStatus"
            ],
            "properties": {
              "componentStatus": {
                "type": "string",
                "enum": [
                  "passed",
                  "partial",
                  "pending",
                  "failed"
                ]
              },
              "realResourceStatus": {
                "type": "string",
                "enum": [
                  "qualified",
                  "pending",
                  "failed",
                  "not_applicable"
                ]
              }
            },
            "additionalProperties": false
          }
        },
        "additionalProperties": false
      },
...
    },
    "responses": {
      "Problem": {
        "description": "Mapped existing SMPP error",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetails"
            }
          }
        }
      }
    },
    "parameters": {
      "CorrelationIdHeader": {
        "name": "X-Correlation-ID",
        "in": "header",
        "required": false,
        "description": "Optional trace correlation identifier. The server may generate one when absent.",
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 128,
          "pattern": "^[A-Za-z0-9._:-]+$"
        }
      },
...
    }
  },
...
}